requirements/base.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
python-decouple==3.0
static3==0.6.1
coverage==4.0.3
django-extensions==1.6.1
django-test-without-migrations==0.4
requests==2.9.1
djangorestframework==3.3.3
pytz
